test(reborn): consolidate gate-refresh scenario into webui_v2_product_api bin
Move approval_gate_rediscovered_and_resolved_after_refresh into the
existing W5-WEBUI-API-1 product-API test binary instead of registering
a second one-test bin for the same hand-built RebornServices/WebUI
facade setup. Test content is unchanged.
