+++ Cargo.toml
@@ -561,10 +561,6 @@ name = "reborn_integration_web_access"
 path = "tests/integration/web_access.rs"
 
 [[test]]
-name = "reborn_integration_webui_v2_gate_refresh"
-path = "tests/integration/webui_v2_gate_refresh.rs"
-
-[[test]]
 name = "reborn_integration_webui_v2_product_api"
 path = "tests/integration/webui_v2_product_api.rs"
